requirements.txt
google-genai
python-dotenv
markdown
supabase
groq
